test : CommentRepositoryTest 테스트 클래스 생성
IssueRepositoryTest에 있던 코멘트 관련 테스트 이동
